#114: green — transition matrix + service + PATCH endpoint
Implements FlightProcessState (8 states), FlightTransitionMatrix (table-driven
port of FlightService.cs:1380-1440), FlightStateTransitionService with
optimistic-locking + audit, and PATCH /api/v1/flights/{id}/process-state.

DeliveryBooked is terminal; illegal transitions raise
IllegalFlightTransitionException (mapped to 409). V12 adds @Version column.

OpenAPI snapshot refreshed for the new endpoint + schemas.

Boyscout: pulls in the handoff skill (.claude/skills/handoff/).

contract.

Cursor error handling: previously the message-prefix sniff in
FlightsExceptionHandler routed IllegalArgumentException to the cursor URI
based on `msg.startsWith("cursor")`, and FlightListCursor.decode let
DateTimeParseException leak through as a 500. Introduced typed
InvalidCursorException; FlightListCursor.decode wraps all four parse failure
paths in it (blank, base64, missing separator, malformed date, malformed
UUID); handler has a dedicated @ExceptionHandler for it. FlightListCursorTest
updated to assert the typed exception.

AircraftAccess: deleted the dead `hasRoleFromAuthorities(Collection,...)`
helper + the now-orphan Optional + GrantedAuthority imports.

OpenAPI snapshot regenerated (FlightUpdateRequest description text changed).

olationException → 400 covers unknown aircraftIds now.
- FlightsControllerIT.unused() static + the FlightId import it propped up.

Code-quality from the panel:

- FlightsExceptionHandler: split INVALID_CURSOR (cursor-decode only) from
  INVALID_REQUEST (runway / coupon / temporal / non-negative aggregate
  invariants). Mislabelling validation errors as cursor errors leaks via
  the RFC 7807 `type` URI.
- FlightsTenantIsolationIT.seedTwoClubs: comment said "Aircraft is now
  tenant-scoped (S-159)"; flip to the S-058 cross-tenant reality.

s doesn't need approval).

ADR 0008 amendment + S-050 / S-058 story-body updates deferred to the
next commit.

ow-up.
- af-textarea atom + af-dialog confirm + routerLink-on-af-button +
  SeedClubIds constant (usability/maintainability nudges) — boyscout
  candidates for the next masterdata story per
  pending-boyscout-followups.md.
- Show-inactive toggle placement nudge — defer to a future filter-strip
  convention story; current placement is the cheap walking-skeleton
  shape.

(CVA) + placeholder;
  drop the manual [value]/(valueChange) shape.
- novalidate on the <form>; markAllAsTouched() when submit hits an
  invalid form; loading-detail gate (matches aircraft-edit pattern);
  drop the allowSignalWrites type-cast (removed in Angular 19+).
- select(id) on entry effect so a stale Person doesn't flash during
  GET round-trip.
- inOtherClubsCount banner: surfaces the privacy signal from the
  Security plan when CLUB_ADMIN opens a Person shared across tenants.

Parity / coverage (improvements):
- New PersonTest: 5 domain unit tests covering joinClub-duplicate-reject,
  joinClub-reactivates-soft-deleted, softDelete-refuses-cross-tenant,
  softDelete-succeeds-when-only-tenant, role-flag-independence.
- New PersonsAuthorizationIT: 3 ITs for 404-not-403 IDOR contract,
  list-excludes-foreign-tenant, anonymous-401.
- New PersonsCrossTenantRideThroughIT: 2 ITs for sacred-cow regression
  guard — Person PK load is cross-tenant; per-tenant list excludes
  foreign-tenant Persons.
- Regenerated openapi.json + FE client picks up LOOKUP_HIT/LOOKUP_MISS.

502 backend tests, 0 failures.

+++ alpenflight/server/src/test/java/ch/alpenflight/flights/domain/FlightTransitionMatrixTest.java
@@ -142,14 +142,14 @@ class FlightTransitionMatrixTest {
     void enum_codes_match_legacy_smallints() {
         // The codes are part of the public contract — the audit payload + e2e
         // parity all key off them. Don't drift.
-        assertThat(FlightProcessState.NOT_PROCESSED.legacyCode()).isEqualTo(0);
-        assertThat(FlightProcessState.INVALID.legacyCode()).isEqualTo(28);
-        assertThat(FlightProcessState.VALID.legacyCode()).isEqualTo(30);
-        assertThat(FlightProcessState.LOCKED.legacyCode()).isEqualTo(40);
-        assertThat(FlightProcessState.DELIVERY_PREPARATION_ERROR.legacyCode()).isEqualTo(45);
-        assertThat(FlightProcessState.DELIVERY_PREPARED.legacyCode()).isEqualTo(50);
-        assertThat(FlightProcessState.DELIVERY_BOOKED.legacyCode()).isEqualTo(60);
-        assertThat(FlightProcessState.EXCLUDED_FROM_DELIVERY_PROCESS.legacyCode()).isEqualTo(99);
+        assertThat(FlightProcessState.NOT_PROCESSED.legacyCode()).isEqualTo((short) 0);
+        assertThat(FlightProcessState.INVALID.legacyCode()).isEqualTo((short) 28);
+        assertThat(FlightProcessState.VALID.legacyCode()).isEqualTo((short) 30);
+        assertThat(FlightProcessState.LOCKED.legacyCode()).isEqualTo((short) 40);
+        assertThat(FlightProcessState.DELIVERY_PREPARATION_ERROR.legacyCode()).isEqualTo((short) 45);
+        assertThat(FlightProcessState.DELIVERY_PREPARED.legacyCode()).isEqualTo((short) 50);
+        assertThat(FlightProcessState.DELIVERY_BOOKED.legacyCode()).isEqualTo((short) 60);
+        assertThat(FlightProcessState.EXCLUDED_FROM_DELIVERY_PROCESS.legacyCode()).isEqualTo((short) 99);
     }
 
     @Test
